FDG-9170 move glitch graphic to its own component
