+++ src/components/page-glitch-graphic/page-glitch-graphic.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import { pxToNumber } from '../../helpers/styles-helper/styles-helper';
 import { breakpointLg } from '../../variables.module.scss';
-import { notFoundGraphic } from './page-error.module.scss';
+import { glitchGraphic } from './page-glitch-graphic.module.scss';
 
-const NotFoundGraphic = () => {
+const GlitchGraphic = () => {
   const png = '/not-found/404gif2.png';
   const gif = '/not-found/404gif.gif';
 
@@ -22,7 +22,7 @@ const NotFoundGraphic = () => {
     };
   }, []);
 
-  return <img className={notFoundGraphic} src={src} alt="404: Page Not Found" role="presentation" />;
+  return <img className={glitchGraphic} src={src} alt="404: Page Not Found" role="presentation" />;
 };
 
-export default NotFoundGraphic;
+export default GlitchGraphic;
